+++ R/spglm.R
@@ -443,7 +443,9 @@ spglm <- function(formula, family, data, spcov_type, xcoord, ycoord, spcov_initi
   # spatial structure can make binomial fits separate far more readily than an
   # ordinary (non-spatial) logistic regression, even with well-behaved
   # covariates; warn when this has happened (see warn_fitted_saturation())
-  warn_fitted_saturation(model_stats$fitted$response, family)
+  if (family == "binomial") {
+    warn_fitted_saturation(expit(model_stats$fitted$link), family)
+  }
 
   # parallel cluster if necessary
   if (data_object$parallel) {